+++ packages/chaindata-provider/src/legacy/index.ts
@@ -1,3 +1,0 @@
-export * from "./Chain"
-export * from "./EvmNetwork"
-export * from "./TalismanChaindataDatabase"