+++ packages/extension-core/src/db/upgrades/2024-01-25-upgradeRemoveSymbolFromNativeTokenId.ts
@@ -1,4 +1,3 @@
-import { DbTokenRates } from "@talismn/token-rates"
 import { Transaction as DbTransaction } from "dexie"
 
 import { LegacyWalletTransaction } from "../../domains/transactions"
@@ -15,19 +14,4 @@ export const upgradeRemoveSymbolFromNativeTokenId = async (tx: DbTransaction) =>
       if (wtx?.tokenId?.includes?.("-evm-native-"))
         wtx.tokenId = wtx.tokenId.replace(/-evm-native-.+$/, "-evm-native")
     })
-
-  const uniqueTokenRates = new Map(
-    (await tx.table<DbTokenRates, string>("tokenRates").toArray()).map((tokenRate) => {
-      if (tokenRate?.tokenId?.includes?.("-substrate-native-"))
-        tokenRate.tokenId = tokenRate.tokenId.replace(/-substrate-native-.+$/, "-substrate-native")
-
-      if (tokenRate?.tokenId?.includes?.("-evm-native-"))
-        tokenRate.tokenId = tokenRate.tokenId.replace(/-evm-native-.+$/, "-evm-native")
-
-      return [tokenRate.tokenId, tokenRate]
-    }),
-  )
-
-  await tx.table<DbTokenRates, string>("tokenRates").clear()
-  await tx.table<DbTokenRates, string>("tokenRates").bulkPut([...uniqueTokenRates.values()])
 }